pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "WhiteoutPDF"
version = "0.1.0"
description = "A simple WhiteoutPDF by Abdullah Adeeb"
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
authors = [
    { name = "Abdullah Adeeb", email = "abdullah@abdullahadeeb.xyz" },
]
keywords = ["WhiteoutPDF", "cli"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
]

dependencies = [
    "pdf2image",
    "Pillow",
]

[project.scripts]
WhiteoutPDF = "WhiteoutPDF.main:main"

[project.urls]
"Repository" = "https://github.com/abdullahadeebx/WhiteoutPDF"
"Bug Tracker" = "https://github.com/abdullahadeebx/WhiteoutPDF/issues"
"Homepage" = "https://www.abdullahadeeb.xyz"
